requirements.txt
numpy
pyodide-py==0.27.7
